pyproject.toml
[project]
name = "desire-hunter"
version = "2.0.0"
description = "Web空間から欲求に最適な製品を自律探索・評価するシステム"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core
    "pydantic>=2.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # Brain - LLM
    "openai>=1.40.0",

    # Eyes & Hands - Web操作
    "firecrawl-py>=1.0.0",
    "requests>=2.31.0",

    # Resilience - エラー耐性
    "tenacity>=8.2.0",
    "ratelimit>=2.2.1",

    # Memory - データ永続化
    "gspread>=6.0.0",
    "oauth2client>=4.1.3",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]
